Big design changes, working pause, preparations for the battle system, a lot of bug fixes and working random ambient system

+++ Space Shooter Vita/Assets/Scripts/Gameplay/ActiveObjectStats.cs
@@ -7,7 +7,7 @@ public class ActiveObjectStats : MonoBehaviour
 	public int objectDamage;
 	public int objectLowerRandomDamage;
 	public int objectHigherRandomDamage;
-	public int objectArmour;
+	public float objectArmour;
 	public int objectMaxHp;
 	public int objectXpGiven;
 }
